Extraer UI de conectividad y sincronización de terminal.js
Tercer paso de la descomposición de terminal.js: setOffline(),
updateIdleSyncStatus(), refreshIdleSyncStatus() y refreshLastSyncLabel()
se mueven a terminal/sync-status-ui.js. Paridad con mark/sync-status-ui.js
(mismo texto/prioridad: conflictos > pendientes > sincronizado), pero no
es el mismo módulo — acá hay además un chequeo previo de "sin empleados
cacheados" (más grave: un terminal sin empleados no puede identificar a
nadie) y un indicador de conectividad de header que mark.js no tiene.

countCachedEmployees/countPendingEvents/countConflictEvents se eliminan
del import de terminal.js (solo los usaba refreshIdleSyncStatus, que ya
vive en el nuevo módulo); getMeta se mantiene porque terminal.js lo
sigue usando directamente en otros puntos (provisión, migración legacy).

Refactor puro, sin cambios de comportamiento.
